requirements.txt
pandas==2.2.2
pillow==10.3.0
jupyter==1.0.0
pillow_heif==0.4.1